refactor(traces): extract TraceComparisonView from command registration
