+++ src/cli/commands/traces/command.tsx
@@ -2,10 +2,10 @@ import { serializeResult } from '../../../lib';
 import { COMMAND_DESCRIPTIONS } from '../../constants';
 import { getErrorMessage } from '../../errors';
 import { loadDeployedProjectConfig } from '../../operations/resolve-agent';
-import type { MetricDelta, TraceComparisonDeltas } from '../../operations/traces';
 import { withCommandRunTelemetry } from '../../telemetry/cli-command-run.js';
 import { getProjectRootMismatch, projectExists, requireProject } from '../../tui/guards';
 import { handleTracesCompare, handleTracesGet, handleTracesList } from './action';
+import { TraceComparisonView } from './comparison-view';
 import type { TracesCompareOptions, TracesGetOptions, TracesListOptions } from './types';
 import type { Command } from '@commander-js/extra-typings';
 import { Box, Text, render } from 'ink';
@@ -18,45 +18,6 @@ function formatTimestamp(ts: string): string {
     .toISOString()
     .replace('T', ' ')
     .replace(/\.\d+Z$/, 'Z');
-}
-
-function formatSeconds(value: number): string {
-  return `${(value / 1000).toFixed(2)}s`;
-}
-
-function formatCount(value: number): string {
-  return String(value);
-}
-
-function formatTokens(value: number): string {
-  return value.toLocaleString('en-US');
-}
-
-function formatSide(value: number | undefined, format: (value: number) => string): string {
-  return value === undefined ? '-' : format(value);
-}
-
-function formatDelta(entry: MetricDelta, format: (value: number) => string): string {
-  if (entry.delta === undefined) return '-';
-  const sign = entry.delta > 0 ? '+' : '';
-  const percent =
-    entry.deltaPercent === null || entry.deltaPercent === undefined
-      ? ''
-      : ` (${entry.deltaPercent > 0 ? '+' : ''}${entry.deltaPercent.toFixed(1)}%)`;
-  return `${sign}${format(entry.delta)}${percent}`;
-}
-
-function buildComparisonRows(deltas: TraceComparisonDeltas) {
-  return [
-    { label: 'End-to-end latency', entry: deltas.endToEndMs, format: formatSeconds },
-    { label: 'LLM latency', entry: deltas.llmMs, format: formatSeconds },
-    { label: 'Tool latency', entry: deltas.toolMs, format: formatSeconds },
-    { label: 'LLM calls', entry: deltas.llmCalls, format: formatCount },
-    { label: 'Tool calls', entry: deltas.toolCalls, format: formatCount },
-    { label: 'Input tokens', entry: deltas.inputTokens, format: formatTokens },
-    { label: 'Output tokens', entry: deltas.outputTokens, format: formatTokens },
-    { label: 'Total tokens', entry: deltas.totalTokens, format: formatTokens },
-  ].filter(row => row.entry.baseline !== undefined || row.entry.candidate !== undefined);
 }
 
 function requireTracesProject(json?: boolean): boolean {
@@ -259,52 +220,14 @@ export const registerTraces = (program: Command) => {
           return;
         }
 
-        const rows = buildComparisonRows(result.deltas);
         render(
-          <Box flexDirection="column">
-            <Text bold>
-              Trace comparison: baseline {result.baseline.traceId} → candidate {result.candidate.traceId}
-            </Text>
-            <Text> </Text>
-            <Box>
-              <Box width={22}>
-                <Text bold>Metric</Text>
-              </Box>
-              <Box width={14} justifyContent="flex-end">
-                <Text bold>Baseline</Text>
-              </Box>
-              <Box width={14} justifyContent="flex-end">
-                <Text bold>Candidate</Text>
-              </Box>
-              <Box width={24} justifyContent="flex-end">
-                <Text bold>Delta</Text>
-              </Box>
-            </Box>
-            {rows.map(row => (
-              <Box key={row.label}>
-                <Box width={22}>
-                  <Text>{row.label}</Text>
-                </Box>
-                <Box width={14} justifyContent="flex-end">
-                  <Text>{formatSide(row.entry.baseline, row.format)}</Text>
-                </Box>
-                <Box width={14} justifyContent="flex-end">
-                  <Text>{formatSide(row.entry.candidate, row.format)}</Text>
-                </Box>
-                <Box width={24} justifyContent="flex-end">
-                  <Text color="cyan">{formatDelta(row.entry, row.format)}</Text>
-                </Box>
-              </Box>
-            ))}
-            {result.warnings.length > 0 && <Text> </Text>}
-            {result.warnings.map(warning => (
-              <Text key={warning} color="yellow">
-                Warning: {warning}
-              </Text>
-            ))}
-            <Text> </Text>
-            {result.consoleUrl && <Text color="gray">Console: {result.consoleUrl}</Text>}
-          </Box>
+          <TraceComparisonView
+            baseline={result.baseline}
+            candidate={result.candidate}
+            deltas={result.deltas}
+            warnings={result.warnings}
+            consoleUrl={result.consoleUrl}
+          />
         );
       } catch (error) {
         if (cliOptions.json) {
